simulation: remove all simulation-related files from backend
They are replaced by the networkDesign. From the socket routes to the
database queries, all simulation files have been deleted.
