+++ packages/transition-backend/src/services/evolutionaryAlgorithm/generation/GenerationLogger.ts
@@ -20,7 +20,8 @@ class GenerationLogger {
     Get a simulation generation and log results to a file
     */
     doLog(generation: Generation): void {
-        const generationNumber = generation.getGenerationNumber();
+        // FIXME Implement or drop? Was commented when changed SimulationRun for Job
+       /* const generationNumber = generation.getGenerationNumber();
         const bestCandidate = generation.getCandidates()[0];
         // TODO, use keepCandidates simulation parameter to log the required number of best candidates.
         const bestCandidateAsString = bestCandidate.toString();
@@ -52,7 +53,7 @@ class GenerationLogger {
                 methods: methods,
                 fitnesses: methods.map((method) => bestResults[method].fitness)
             });
-        }
+        } */
         // TODO: implement json logger
     }
 }
