+++ packages/transition-frontend/src/components/dashboard/TransitionToolbar.tsx
@@ -139,7 +139,6 @@ const Toolbar: React.FunctionComponent<LayoutSectionProps> = (_props) => {
                 pathCollection: serviceLocator.collectionManager.get('paths'),
                 nodeCollection: serviceLocator.collectionManager.get('nodes'),
                 placeCollection: serviceLocator.collectionManager.get('places'),
-                simulationCollection: serviceLocator.collectionManager.get('simulations'),
                 dataSourceCollection: serviceLocator.collectionManager.get('dataSources')
             });
         } catch (error) {
